Code cleanup - changed array size numbers to use DEFINEs - changed some int's to byte's in for loops - updated comment in globals file

+++ RF/RF.cpp
@@ -57,7 +57,7 @@ byte RFClass::RFCheck()
 
 void RFClass::RadionWrite()
 {
-	for (int a=0;a<RF_CHANNELS;a++)
+	for (byte a=0;a<RF_CHANNELS;a++)
 	{
 		SetMode(Radion, RadionChannels[a], a);
 	}
